Remove redundant backend
