--- ExecutionService/dockerfile
@@ -1,9 +1,0 @@
-FROM node:22
-
-WORKDIR /app
-COPY package.json /app
-RUN npm install
-COPY . /app
-
-EXPOSE 3010
-ENTRYPOINT ["npm", "start"]
