--- ExecutionService/package.json
@@ -1,12 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "start": "node app"
-  },
-  "dependencies": {
-    "axios": "^1.7.7",
-    "cors": "^2.8.5",
-    "express": "^4.21.1",
-    "vm2": "^3.9.19"
-  }
-}
